+++ MailChimp/Models/Clicks.cs
@@ -8,16 +8,16 @@ namespace MailChimpWrapper.Models
 	public class Clicks
 	{
 		[JsonPropertyName("clicks_total")]
-		public long ClicksTotal { get; set; }
+		public int ClicksTotal { get; set; }
 
 		[JsonPropertyName("unique_clicks")]
-		public long UniqueClicks { get; set; }
+		public int UniqueClicks { get; set; }
 
 		[JsonPropertyName("unique_subscriber_clicks")]
-		public long UniqueSubscriberClicks { get; set; }
+		public int UniqueSubscriberClicks { get; set; }
 
 		[JsonPropertyName("click_rate")]
-		public long ClickRate { get; set; }
+		public int ClickRate { get; set; }
 
 		[JsonPropertyName("last_click")]
 		public string LastClick { get; set; }
