+++ MailChimp/Models/Constraints.cs
@@ -5,15 +5,15 @@ using System.Text.Json.Serialization;
 
 namespace MailChimpWrapper.Models
 {
-    public class Constraints
-    {
-        [JsonPropertyName("may_create")]
-        public bool MayCreate { get; set; }
+	public class Constraints
+	{
+		[JsonPropertyName("may_create")]
+		public bool MayCreate { get; set; }
 
-        [JsonPropertyName("max_instances")]
-        public long MaxInstances { get; set; }
+		[JsonPropertyName("max_instances")]
+		public int MaxInstances { get; set; }
 
-        [JsonPropertyName("current_total_instances")]
-        public long CurrentTotalInstances { get; set; }
-    }
+		[JsonPropertyName("current_total_instances")]
+		public int CurrentTotalInstances { get; set; }
+	}
 }
