+++ MailChimp/Models/Responses/BatchSubscribeUnsubscribeResponse.cs
@@ -17,13 +17,13 @@ namespace MailChimpWrapper.Models.Responses
 		public MemberError[] Errors { get; set; }
 
 		[JsonPropertyName("total_created")]
-		public long TotalCreated { get; set; }
+		public int TotalCreated { get; set; }
 
 		[JsonPropertyName("total_updated")]
-		public long TotalUpdated { get; set; }
+		public int TotalUpdated { get; set; }
 
 		[JsonPropertyName("error_count")]
-		public long ErrorCount { get; set; }
+		public int ErrorCount { get; set; }
 
 		[JsonPropertyName("_links")]
 		public Link[] Links { get; set; }
