+++ MailChimp/Models/Responses/ListsResponse.cs
@@ -6,26 +6,26 @@ using MailChimpWrapper.Models.Requests;
 
 namespace MailChimpWrapper.Models.Responses
 {
-    /// <summary>
-    /// This is the result of a request:
-    /// <para>
-    /// <see cref="ListsGetRequest"></see>
-    /// </para>
-    /// </summary>
-    public class ListsResponse : IBaseResponse
-    {
-        [JsonPropertyName("lists")]
-        public List<SubscriberList> Lists { get; set; }
+	/// <summary>
+	/// This is the result of a request:
+	/// <para>
+	/// <see cref="ListsGetRequest"></see>
+	/// </para>
+	/// </summary>
+	public class ListsResponse : IBaseResponse
+	{
+		[JsonPropertyName("lists")]
+		public List<SubscriberList> Lists { get; set; }
 
-        [JsonPropertyName("total_items")]
-        public long TotalItems { get; set; }
+		[JsonPropertyName("total_items")]
+		public int TotalItems { get; set; }
 
-        [JsonPropertyName("constraints")]
-        public Constraints Constraints { get; set; }
+		[JsonPropertyName("constraints")]
+		public Constraints Constraints { get; set; }
 
-        [JsonPropertyName("_links")]
-        public Link[] Links { get; set; }
+		[JsonPropertyName("_links")]
+		public Link[] Links { get; set; }
 
-        public int Status { get; set; }
-    }
+		public int Status { get; set; }
+	}
 }
